chore: move to current releases (numpy 2.5, librosa 1.0, demucs 4.1, torch 2.14); MCP SDK notes; slimmer icon
Environments (all verified: 40/40 tests; every benchmark artifact identical
to the previous stack - beat times to 0.0 ms, keys, meter, chords, mode):
- analysis: Python 3.12, numpy 2.5.3, madmom git 27f032e (builds on numpy 2
  + Cython 3), librosa 1.0.0, scipy 1.18, TensorFlow 2.21, basic-pitch
  (--no-deps), crepe (setuptools<81 build constraint)
- stems: demucs 4.1.0 + torch 2.14 (same weights, ~3.5x less CPU)
- chords: Python 3.12, lv-chordia 1.1.0, librosa 1.0, torch 2.14

Patches the upgrade needed:
- librosa 1.0 dropped audioread: lv_chords/detect_meter transcode
  m4a/aac/webm with ffmpeg to temp WAVs (cleaned up at exit)
- demucs 4.1 reads MP3 without the gapless trim (stems 25 ms late): decode
  to WAV first, documented everywhere demucs runs; htdemucs_6s verified
- validate_artifacts --check-cache finds 4.1's Hugging Face cache
- run_bench prints plain floats under numpy 2

MCP: the fork runs unchanged up to mcp 1.12.2; with a small patch (import
MCPServer when present, instructions= kwarg, lock around socket I/O) on the
latest 1.x (1.30.0, tested live). mcp 2.x not supported yet. Documented the
uvx cache trap and the editable-venv setup.

README: requirements reflect the new stack. Pipeline icon: slimmer eighths.

+++ song-analysis/bench/run_bench.py
@@ -163,8 +163,8 @@ def score_song(song, root, tasks):
         got = [l.get("start") for l in rd("lyrics_aligned.json")["lines"]]
         n = min(len(truth), len(got))
         err = [abs(got[i] - truth[i]["start"]) for i in range(n) if got[i] is not None]
-        res["lines"] = dict(n=n, within_0_5s=round(100 * np.mean([e <= 0.5 for e in err]), 1),
-                            within_1s=round(100 * np.mean([e <= 1.0 for e in err]), 1),
+        res["lines"] = dict(n=n, within_0_5s=round(100 * float(np.mean([e <= 0.5 for e in err])), 1),
+                            within_1s=round(100 * float(np.mean([e <= 1.0 for e in err])), 1),
                             median_err_s=round(float(np.median(err)), 2)) if err else None
     return res
